Find all points when a numeric vector crosses some threshold

This function takes a vector of y values and returns the index or x value of every point where the y values cross some threshold y value.

Usage

find_threshold_crosses(
  y,
  x = NULL,
  threshold,
  return = "index",
  return_rising = TRUE,
  return_falling = TRUE,
  return_endpoints = TRUE,
  subset = NULL,
  na.rm = TRUE
)

Arguments

y

Numeric vector of y values in which to identify threshold crossing events

x

Optional numeric vector of corresponding x values

threshold

Threshold y value of interest

return

One of c("index", "x"), determining whether the function will return the index or x value associated with the threshold-crossing event.

If index, it will refer to the data point immediately after the crossing event.

If x, it will use linear interpolation and the data points immediately before and after the threshold-crossing to return the exact x value when the threshold crossing occurred

return_rising

Boolean for whether crossing events where y rises above threshold should be returned

return_falling

Boolean for whether crossing events where y falls below threshold should be returned

return_endpoints

Boolean for whether startpoint should be returned when the startpoint is above threshold and return_rising = TRUE, or when the startpoint is below threshold and return_falling = TRUE

subset

A vector of Boolean values indicating which x and y values should be included (TRUE) or excluded (FALSE).

If return = "index", index will be for the whole vector and not the subset of the vector

na.rm

Boolean whether NA's should be removed before analyzing. If return = 'index', indices will refer to the original y vector *including* NA values

Value

A vector of indices (return = "index") or x values (return = "x") for when y crossed threshold