Import blockdesigns

Function to import block-shaped designs from files and return tidy designs. This function acts as a wrapper to call read_blocks, paste_blocks, trans_block_to_wide, trans_wide_to_tidy, and separate_tidys in one go

Usage

import_blockdesigns(files, block_names = NULL, sep = NULL, ...)

Arguments

files

Vector of filenames (as strings), each of which is a block-shaped designs file. Inputs can be .csv, .xls, or .xlsx

block_names

Vector of names for design elements. These will be the resulting column names in the output data frame. Should be in the same order as files and/or same order as corresponding files themselves. If NULL, file names will be used as column names.

sep

If block design files are already pasted, sep specifies the string separating design elements

If NULL, import_blockdesigns will assume no elements are already pasted together and attempt to find a character not used in the imported files to paste and later separate design elements.

...

Other arguments to pass to read_blocks, paste_blocks, trans_block_to_wide, trans_wide_to_tidy, or separate_tidy.

See Details for more information

Value

A tidy-shaped data.frame containing the design information from files

Details

Common arguments that you may want to provide via ... include:

startrow, endrow, startcol, endcol, sheet - specifying the location of design information inside files to read_blocks

wellnames_sep - specifying what character (or "" for none) should be used when pasting together the rownames and column names. Note that this should be chosen to match the well names in your measures.

Note that import_blockdesigns cannot currently handle metadata specified via the metadata argument of read_blocks

If you find yourself needing more control, you can run the steps manually, first reading with read_blocks, pasting as needed with paste_blocks, transforming to tidy with trans_block_to_wide and trans_wide_to_tidy, and separating as needed with separate_tidys.